runtime: split established instance responsibilities
Why:
- the staged transaction made instance state, close ownership, and import attachment responsibilities explicit, but instantiate.go still mixed all four concerns

What:
- keep construction stages in instantiate.go
- move live instance state, lifecycle release, and import attachment/root-retention code into cohesive files
- document the source boundaries in ARCHITECTURE.md

Proof:
- go test ./src/wago -count=1
- instantiate.go reduced from 1451 to 966 lines with no behavior change

Docs:
- ARCHITECTURE.md records the responsibility-based file layout
