+++ ARCHITECTURE.md
@@ -86,6 +86,10 @@ function index. `Instantiate` turns a `*Compiled` into a runnable `*Instance`.
 
 ```
 src/wago/                         public API implementation (package wago)
+  instantiate.go                  staged instance-construction transaction
+  instance.go                     live instance state and native-visible handles
+  instance_lifecycle.go           close/physical-release ownership transfer
+  import_attachments.go           imported owner attachment and root retention
 wago.go                           generated root facade (re-exports src/wago)
 internal/genfacade/               generator for wago.go (+ up-to-date test)
 cli/wago/                         CLI entry point and command implementation
