+++ src/wago/instance_lifecycle.go
@@ -1,0 +1,150 @@
+package wago
+
+import (
+	"errors"
+	"fmt"
+
+	"github.com/wago-org/wago/src/core/runtime"
+)
+
+// Close releases the instance's mapped code, engine, and (if instance-owned) its
+// memory. An imported memory is left for the host to Close. Close is idempotent.
+// Concurrent callers wait for the active close operation and receive its same
+// completed, possibly aggregated, error result.
+func (in *Instance) Close() (err error) {
+	if in == nil {
+		return nil
+	}
+	state, owner := in.beginClose()
+	if !owner {
+		<-state.done
+		return state.result
+	}
+	defer func() {
+		if recovered := recover(); recovered != nil {
+			if panicErr, ok := recovered.(error); ok {
+				err = joinPrimary(err, fmt.Errorf("wago: instance close panicked: %w", panicErr))
+			} else {
+				err = joinPrimary(err, fmt.Errorf("wago: instance close panicked: %v", recovered))
+			}
+		}
+		state.result = err
+		close(state.done)
+	}()
+	return in.closeOnce()
+}
+
+func (in *Instance) beginClose() (*instanceCloseState, bool) {
+	state := in.ensurePluginState()
+	if active := state.close.Load(); active != nil {
+		return active, false
+	}
+	candidate := &instanceCloseState{done: make(chan struct{})}
+	if state.close.CompareAndSwap(nil, candidate) {
+		return candidate, true
+	}
+	return state.close.Load(), false
+}
+
+func (in *Instance) closeOnce() error {
+	var errs []error
+	appendStep := func(phase string, fn func()) {
+		if err := callHookSafely(phase, fn); err != nil {
+			errs = append(errs, err)
+		}
+	}
+
+	in.lifeMu.Lock()
+	alreadyClosed := in.closed
+	in.lifeMu.Unlock()
+	if alreadyClosed {
+		return nil
+	}
+
+	if in.rt != nil {
+		for i := len(in.rt.hooks.internalBeforeClose) - 1; i >= 0; i-- {
+			fn := in.rt.hooks.internalBeforeClose[i]
+			appendStep("internal BeforeClose", func() { fn(in) })
+		}
+	}
+
+	var hctx *InstanceContext
+	if in.rt != nil && (len(in.rt.hooks.beforeClose) != 0 || len(in.rt.hooks.afterClose) != 0) {
+		hctx = &InstanceContext{
+			Runtime: in.rt, Compiled: in.c, Instance: in, Origin: in.instantiateOrigin(), Metadata: map[string]any{},
+		}
+		for i := len(in.rt.hooks.beforeClose) - 1; i >= 0; i-- {
+			fn := in.rt.hooks.beforeClose[i]
+			appendStep("BeforeClose", func() { fn(hctx) })
+		}
+	}
+
+	// Before marking the instance closed, transfer producer roots to any imported
+	// funcref table or global that still holds a local funcref this instance wrote
+	// (via table.set/fill/grow/init or global.set). The descriptor embeds this
+	// instance's code pointer and home linear-memory address, so it must outlive
+	// the write for other importers that later read it. retainResourceRoot refuses
+	// a closed instance, so this runs before in.closed is set; the container drops
+	// the root when the descriptor is overwritten or the container closes.
+	appendStep("retain imported table roots", func() { retainProducerRootsInImportedTables(in) })
+	appendStep("retain imported global roots", func() { retainProducerRootsInImportedGlobals(in) })
+
+	in.lifeMu.Lock()
+	in.closed = true
+	shouldRelease := in.resourceRefs == 0
+	store := in.refStore
+	in.lifeMu.Unlock()
+
+	appendStep("detach imported host functions", func() { detachImportedHostFuncRefs(in) })
+	appendStep("detach imported globals", func() { detachImportedGlobals(in) })
+	appendStep("detach imported tables", func() { detachImportedTables(in) })
+	if store != nil {
+		appendStep("close reference store instance", func() { store.instanceClosed(in) })
+	}
+	if shouldRelease {
+		appendStep("release instance resources", in.releaseResources)
+	}
+	if hctx != nil {
+		for i := len(in.rt.hooks.afterClose) - 1; i >= 0; i-- {
+			fn := in.rt.hooks.afterClose[i]
+			appendStep("AfterClose", func() { fn(hctx) })
+		}
+	}
+	return errors.Join(errs...)
+}
+
+func (in *Instance) releaseResources() {
+	in.lifeMu.Lock()
+	if in.resourcesClosed {
+		in.lifeMu.Unlock()
+		return
+	}
+	in.resourcesClosed = true
+	in.lifeMu.Unlock()
+
+	if in.gc != nil {
+		in.gc.Close()
+	}
+	for table := in.table; table != nil; table = table.next {
+		table.releaseRetainedInstances()
+	}
+	if in.thunkMem != nil {
+		runtime.Unmap(in.thunkMem)
+		in.thunkMem = nil
+	}
+	in.c.releaseCode()
+	runtime.ReleaseArena(in.ar)
+	if in.ownsMem {
+		if in.memory != nil {
+			in.memory.ownerClosed()
+		}
+		runtime.ReleaseJobMemory(in.jm)
+	} else if in.memory != nil {
+		in.memory.detachImporter()
+	}
+	runtime.ReleaseEngine(in.eng)
+}
+
+// Memory returns the instance's linear-memory object (instance-owned or the
+// host-imported one). Use Memory().Bytes() for the zero-copy byte view.
+func (in *Instance) Memory() *Memory { return in.memory }
