+++ src/wago/instantiate.go
@@ -4,82 +4,11 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
-	"sync"
-	"sync/atomic"
 	"unsafe"
 
 	"github.com/wago-org/wago/src/core/runtime"
 	"github.com/wago-org/wago/src/core/runtime/gc"
 )
-
-// offHeapPtr reinterprets a known off-heap address — JIT arena / table-descriptor
-// memory, kept live by arena/table ownership and never on the Go heap — as an
-// unsafe.Pointer. Routing through *uintptr avoids a direct uintptr→unsafe.Pointer
-// conversion, which go vet's unsafeptr pass flags (it cannot prove the target is
-// non-heap). Use ONLY for addresses into that off-heap memory; there is no
-// live-pointer hazard there.
-func offHeapPtr(addr uintptr) unsafe.Pointer {
-	return *(*unsafe.Pointer)(unsafe.Pointer(&addr))
-}
-
-// Instance is ready for repeated Invoke calls.
-type Instance struct {
-	c                      *Compiled
-	eng                    *runtime.Engine
-	jm                     *runtime.JobMemory
-	memory                 *Memory // the memory object (owned or host-imported)
-	ar                     *runtime.Arena
-	base                   uintptr
-	hosts                  map[string]HostFunc
-	imports                Imports // the imports as provided to Instantiate
-	hostLog                []byte
-	ctrl                   []byte                              // sync host-call control frame (nil in async mode)
-	syncHosts              []HostFunc                          // per import-func-index host, sync mode only
-	hostCall               runtime.HostCall                    // per-instance sync host dispatcher, allocated once
-	pluginState            atomic.Pointer[instancePluginState] // allocated only after privileged instance services activate
-	globals                []byte                              // pointer table handed to JIT code
-	globalCells            []*Global
-	table                  *Table        // lazily created importer-owned local export-handle chain
-	tableDescPtr           uintptr       // local/imported descriptor address; arena/table ownership keeps it live
-	tableDescLen           int           // descriptor byte length for safe slice reconstruction
-	funcRefDescs           []byte        // canonical funcref descriptor handles for this instance's function index space
-	passiveDataDesc        []byte        // per-instance data-segment descriptors; active slots start dropped
-	thunkMem               []byte        // executable mapping for host-func-in-table log thunks (nil if none)
-	gc                     *gc.Collector // nil for modules with no Wasm GC descriptors/runtime use
-	serArgs, results, trap []byte
-	resultVals             []uint64       // reusable Invoke result buffer (valid until the next call)
-	ic                     [4]invokeCache // tiny fixed export resolution cache
-	icNext                 uint8          // round-robin replacement cursor
-	refStore               *referenceStore
-	lifeMu                 sync.Mutex
-	resourceRefs           int
-	closed                 bool // logical close; retained references may defer physical release
-	resourcesClosed        bool
-	ownsMem                bool    // false when memory is host-imported (don't close it)
-	syncMode               bool    // true when host imports use the synchronous re-entry protocol
-	nativeControlShared    bool    // entered from another instance; prepared control fields may be overwritten
-	nativeContext          uintptr // arena-backed context bytes rebound before every native entry
-
-	// rt is set when the instance is created through Runtime.Instantiate, so
-	// Instance.Call and Instance.Close can fire lifecycle hooks. It is nil for
-	// low-level package-level Instantiate, which stays hook-free.
-	rt *Runtime
-}
-
-// invokeCache memoizes per-export work so hot Invoke loops skip the exports map
-// probe and the fat ValType width comparisons on every call. Instance keeps a
-// few fixed slots because real AS loops commonly interleave the business export
-// with __collect, __pin, or paired request/response exports.
-type invokeCache struct {
-	export            string
-	valid             bool
-	li                int // local index, or -1-import index for an InstanceExport re-export
-	paramSlots        int
-	resultSlots       int
-	hasFuncRefParams  bool
-	hasFuncRefResults bool
-	resultWide        []bool // one entry per returned uint64 slot; false means read low 32 bits
-}
 
 // InstantiateOptions configures instance creation from a *Compiled. When
 // instantiating from a *Snapshot both fields are ignored: the snapshot carries
@@ -1035,417 +964,3 @@ func buildHostFuncThunks(c *Compiled, imports Imports, syncMode bool) (map[uint3
 	}
 	return addr, mem, nil
 }
-
-// Close releases the instance's mapped code, engine, and (if instance-owned) its
-// memory. An imported memory is left for the host to Close. Close is idempotent.
-// Concurrent callers wait for the active close operation and receive its same
-// completed, possibly aggregated, error result.
-func (in *Instance) Close() (err error) {
-	if in == nil {
-		return nil
-	}
-	state, owner := in.beginClose()
-	if !owner {
-		<-state.done
-		return state.result
-	}
-	defer func() {
-		if recovered := recover(); recovered != nil {
-			if panicErr, ok := recovered.(error); ok {
-				err = joinPrimary(err, fmt.Errorf("wago: instance close panicked: %w", panicErr))
-			} else {
-				err = joinPrimary(err, fmt.Errorf("wago: instance close panicked: %v", recovered))
-			}
-		}
-		state.result = err
-		close(state.done)
-	}()
-	return in.closeOnce()
-}
-
-func (in *Instance) beginClose() (*instanceCloseState, bool) {
-	state := in.ensurePluginState()
-	if active := state.close.Load(); active != nil {
-		return active, false
-	}
-	candidate := &instanceCloseState{done: make(chan struct{})}
-	if state.close.CompareAndSwap(nil, candidate) {
-		return candidate, true
-	}
-	return state.close.Load(), false
-}
-
-func (in *Instance) closeOnce() error {
-	var errs []error
-	appendStep := func(phase string, fn func()) {
-		if err := callHookSafely(phase, fn); err != nil {
-			errs = append(errs, err)
-		}
-	}
-
-	in.lifeMu.Lock()
-	alreadyClosed := in.closed
-	in.lifeMu.Unlock()
-	if alreadyClosed {
-		return nil
-	}
-
-	if in.rt != nil {
-		for i := len(in.rt.hooks.internalBeforeClose) - 1; i >= 0; i-- {
-			fn := in.rt.hooks.internalBeforeClose[i]
-			appendStep("internal BeforeClose", func() { fn(in) })
-		}
-	}
-
-	var hctx *InstanceContext
-	if in.rt != nil && (len(in.rt.hooks.beforeClose) != 0 || len(in.rt.hooks.afterClose) != 0) {
-		hctx = &InstanceContext{
-			Runtime: in.rt, Compiled: in.c, Instance: in, Origin: in.instantiateOrigin(), Metadata: map[string]any{},
-		}
-		for i := len(in.rt.hooks.beforeClose) - 1; i >= 0; i-- {
-			fn := in.rt.hooks.beforeClose[i]
-			appendStep("BeforeClose", func() { fn(hctx) })
-		}
-	}
-
-	// Before marking the instance closed, transfer producer roots to any imported
-	// funcref table or global that still holds a local funcref this instance wrote
-	// (via table.set/fill/grow/init or global.set). The descriptor embeds this
-	// instance's code pointer and home linear-memory address, so it must outlive
-	// the write for other importers that later read it. retainResourceRoot refuses
-	// a closed instance, so this runs before in.closed is set; the container drops
-	// the root when the descriptor is overwritten or the container closes.
-	appendStep("retain imported table roots", func() { retainProducerRootsInImportedTables(in) })
-	appendStep("retain imported global roots", func() { retainProducerRootsInImportedGlobals(in) })
-
-	in.lifeMu.Lock()
-	in.closed = true
-	shouldRelease := in.resourceRefs == 0
-	store := in.refStore
-	in.lifeMu.Unlock()
-
-	appendStep("detach imported host functions", func() { detachImportedHostFuncRefs(in) })
-	appendStep("detach imported globals", func() { detachImportedGlobals(in) })
-	appendStep("detach imported tables", func() { detachImportedTables(in) })
-	if store != nil {
-		appendStep("close reference store instance", func() { store.instanceClosed(in) })
-	}
-	if shouldRelease {
-		appendStep("release instance resources", in.releaseResources)
-	}
-	if hctx != nil {
-		for i := len(in.rt.hooks.afterClose) - 1; i >= 0; i-- {
-			fn := in.rt.hooks.afterClose[i]
-			appendStep("AfterClose", func() { fn(hctx) })
-		}
-	}
-	return errors.Join(errs...)
-}
-
-func (in *Instance) releaseResources() {
-	in.lifeMu.Lock()
-	if in.resourcesClosed {
-		in.lifeMu.Unlock()
-		return
-	}
-	in.resourcesClosed = true
-	in.lifeMu.Unlock()
-
-	if in.gc != nil {
-		in.gc.Close()
-	}
-	for table := in.table; table != nil; table = table.next {
-		table.releaseRetainedInstances()
-	}
-	if in.thunkMem != nil {
-		runtime.Unmap(in.thunkMem)
-		in.thunkMem = nil
-	}
-	in.c.releaseCode()
-	runtime.ReleaseArena(in.ar)
-	if in.ownsMem {
-		if in.memory != nil {
-			in.memory.ownerClosed()
-		}
-		runtime.ReleaseJobMemory(in.jm)
-	} else if in.memory != nil {
-		in.memory.detachImporter()
-	}
-	runtime.ReleaseEngine(in.eng)
-}
-
-// Memory returns the instance's linear-memory object (instance-owned or the
-// host-imported one). Use Memory().Bytes() for the zero-copy byte view.
-func (in *Instance) Memory() *Memory { return in.memory }
-
-// importDedup is an insertion-ordered set of distinct comparable values — the
-// engine uses it for import owner pointers (host funcrefs, reference globals,
-// imported tables). A small inline array keeps the common case (a handful of
-// imports) allocation-free; overflow spills to a slice. The zero value is ready
-// to use.
-type importDedup[T comparable] struct {
-	inline [4]T
-	n      int
-	extra  []T
-}
-
-func (d *importDedup[T]) contains(v T) bool {
-	for i := 0; i < d.n && i < len(d.inline); i++ {
-		if d.inline[i] == v {
-			return true
-		}
-	}
-	for _, e := range d.extra {
-		if e == v {
-			return true
-		}
-	}
-	return false
-}
-
-// push appends v unconditionally; callers needing dedup use add or guard with
-// contains first.
-func (d *importDedup[T]) push(v T) {
-	if d.n < len(d.inline) {
-		d.inline[d.n] = v
-	} else {
-		d.extra = append(d.extra, v)
-	}
-	d.n++
-}
-
-// add inserts v if absent and reports whether it was newly inserted.
-func (d *importDedup[T]) add(v T) bool {
-	if d.contains(v) {
-		return false
-	}
-	d.push(v)
-	return true
-}
-
-// each calls fn for every distinct element in insertion order.
-func (d *importDedup[T]) each(fn func(T)) {
-	inlineCount := d.n
-	if inlineCount > len(d.inline) {
-		inlineCount = len(d.inline)
-	}
-	for i := 0; i < inlineCount; i++ {
-		fn(d.inline[i])
-	}
-	for _, e := range d.extra {
-		fn(e)
-	}
-}
-
-// reset empties the set, clearing the inline array so it retains no references.
-func (d *importDedup[T]) reset() {
-	inlineCount := d.n
-	if inlineCount > len(d.inline) {
-		inlineCount = len(d.inline)
-	}
-	var zero T
-	for i := 0; i < inlineCount; i++ {
-		d.inline[i] = zero
-	}
-	d.n = 0
-	d.extra = nil
-}
-
-type hostFuncRefAttachments struct {
-	set importDedup[*HostFuncRef]
-}
-
-func (a *hostFuncRefAttachments) attach(owner *HostFuncRef, store *referenceStore, sig FuncSig) error {
-	if owner == nil {
-		return fmt.Errorf("host funcref owner is nil")
-	}
-	if a.set.contains(owner) {
-		return owner.validateImport(store, sig)
-	}
-	if err := owner.attachImporter(store, sig); err != nil {
-		return err
-	}
-	a.set.push(owner)
-	return nil
-}
-
-func (a *hostFuncRefAttachments) detachAll() {
-	a.set.each((*HostFuncRef).detachImporter)
-	a.set.reset()
-}
-
-func detachImportedHostFuncRefs(in *Instance) {
-	if in == nil || in.c == nil {
-		return
-	}
-	var seen importDedup[*HostFuncRef]
-	for _, key := range in.c.Imports {
-		owner, ok := in.imports[key].(*HostFuncRef)
-		if !ok || owner == nil {
-			continue
-		}
-		if seen.add(owner) {
-			owner.detachImporter()
-		}
-	}
-}
-
-type globalImportAttachments struct {
-	set importDedup[*Global]
-}
-
-func (a *globalImportAttachments) attach(global *Global, store *referenceStore) error {
-	if global == nil {
-		return fmt.Errorf("reference global is nil")
-	}
-	if a.set.contains(global) {
-		return global.validateReferenceImport(store)
-	}
-	if err := global.attachReferenceImporter(store); err != nil {
-		return err
-	}
-	a.set.push(global)
-	return nil
-}
-
-func (a *globalImportAttachments) detachAll() {
-	a.set.each((*Global).detachReferenceImporter)
-	a.set.reset()
-}
-
-func detachImportedGlobals(in *Instance) {
-	if in == nil || in.c == nil {
-		return
-	}
-	var seen importDedup[*Global]
-	for _, imp := range in.c.GlobalImports {
-		if !isReferenceValType(imp.Type) {
-			continue
-		}
-		provided, ok := in.imports.global(imp.Module + "." + imp.Name)
-		if !ok || provided.Global == nil {
-			continue
-		}
-		if seen.add(provided.Global) {
-			provided.Global.detachReferenceImporter()
-		}
-	}
-}
-
-func retainProducerRootsInImportedGlobals(in *Instance) bool {
-	if in == nil || in.c == nil {
-		return false
-	}
-	retained := false
-	var seen importDedup[*Global]
-	for _, imp := range in.c.GlobalImports {
-		if imp.Type != ValFuncRef {
-			continue
-		}
-		provided, ok := in.imports.global(imp.Module + "." + imp.Name)
-		if !ok || provided.Global == nil {
-			continue
-		}
-		if !seen.add(provided.Global) {
-			continue
-		}
-		if provided.Global.retainProducerInstance(in) {
-			retained = true
-		}
-	}
-	return retained
-}
-
-type tableImportAttachments struct {
-	set importDedup[*Table]
-}
-
-func (a *tableImportAttachments) attach(table *Table, elementType ValType, store *referenceStore) error {
-	if err := table.validateImport(elementType, store); err != nil {
-		return err
-	}
-	if a.set.contains(table) {
-		return nil
-	}
-	if err := table.attachImporter(elementType, store); err != nil {
-		return err
-	}
-	a.set.push(table)
-	return nil
-}
-
-func (a *tableImportAttachments) detachAll() {
-	a.set.each((*Table).detachImporter)
-	a.set.reset()
-}
-
-func detachImportedTables(in *Instance) {
-	if in == nil || in.c == nil {
-		return
-	}
-	var seen importDedup[*Table]
-	for tableIndex := 0; tableIndex < in.c.tableImportCount(); tableIndex++ {
-		def, _ := in.c.tableImportAt(tableIndex)
-		table, ok := in.imports.table(def.Key)
-		if !ok || table == nil {
-			continue
-		}
-		if seen.add(table) {
-			table.detachImporter()
-		}
-	}
-}
-
-func retainProducerRootsInImportedTables(in *Instance) bool {
-	if in == nil || in.c == nil {
-		return false
-	}
-	retained := false
-	for tableIndex := 0; tableIndex < in.c.tableImportCount(); tableIndex++ {
-		def, _ := in.c.tableImportAt(tableIndex)
-		table, ok := in.imports.table(def.Key)
-		if ok && table.retainProducerInstance(in) {
-			retained = true
-		}
-	}
-	return retained
-}
-
-func (in *Instance) tableDescriptor(index int) []byte {
-	if in == nil || in.c == nil || index < 0 || index >= in.c.tableCount() {
-		return nil
-	}
-	if importDef, imported := in.c.tableImportAt(index); imported {
-		table, ok := in.imports.table(importDef.Key)
-		if !ok || len(table.desc) < 8 {
-			return nil
-		}
-		return table.desc
-	}
-	if index == 0 {
-		if in.tableDescPtr == 0 || in.tableDescLen <= 0 {
-			return nil
-		}
-		return unsafe.Slice((*byte)(offHeapPtr(in.tableDescPtr)), in.tableDescLen)
-	}
-	dirPtr := in.jm.TableDirPtr()
-	if dirPtr == 0 {
-		return nil
-	}
-	dir := unsafe.Slice((*byte)(offHeapPtr(dirPtr)), 8*in.c.tableCount())
-	descPtr := uintptr(binary.LittleEndian.Uint64(dir[index*8:]))
-	if descPtr == 0 {
-		return nil
-	}
-	def := in.c.tableDef(index)
-	capacity := def.Max
-	if capacity == 0 {
-		capacity = def.Size
-	}
-	return unsafe.Slice((*byte)(offHeapPtr(descPtr)), 8+capacity*in.c.tableEntryBytes(index))
-}
-
-// Imports returns the imports map this instance was created with, for retrieving
-// imported objects (e.g. a *Memory or *Global) by "module.name" key. The map is
-// the one passed to Instantiate; do not mutate it.
-func (in *Instance) Imports() Imports { return in.imports }
